requirements.txt
python-telegram-bot==20.6.0
python-dotenv==1.0.0
psycopg2==2.9.9